backtracking problem
 we need to write a recursive backtracking C++ program that divides the 
given message string to individual meaningful words. A meaningful word is one that exists in the 
English dictionary. The program should try to form words starting from the first letter. If it finds the 
word in the dictionary, it calls itself on the remaining text. The program stops as soon as it can divide 
the message to a useful sentence with all its words in the dictionary. If no solution is found the program 
will do exhaustive recursion of all possible solutions and then fail. For example, if the message is 
"Iwearmyshoes", it will try to divide as follows:
 "I" is a word in dictionary, then repeat on the remaining letters.
 Try "w" which is not a word, then try "we" which is a word and repeat on the rest.
 Try "a" which is a word, then repeat on the rest. (So far we have "I we a")
 Then we try "r" which is not a word and "rm" which is not a word. We can keep trying bigger words 
like "rmy" and "rmys" or if we find no word starts with "rm" we can stop.
 We backtrack and try "ar" which is word and then repeat on the rest "myshoes"
 And so on. The first valid solution will be "I we ar my shoes".